refactor: move CheckboxRow and IconTextButton to shared components
