+++ app/components/card-screen/index.ts
@@ -5,7 +5,6 @@ export * from "./card-balance-section"
 export * from "./card-transaction-item"
 export * from "./card-transactions-section"
 export * from "./contact-support-row"
-export * from "./icon-text-button"
 export * from "./info-card"
 export * from "./info-row"
 export * from "./info-section"
@@ -17,7 +16,6 @@ export * from "./statement-item"
 export * from "./status-badge"
 export * from "./switch-row"
 export * from "./selectable-option-row"
-export * from "./checkbox-row"
 export * from "./shipping-address-form"
 export * from "./card-status-layout"
 export * from "./dropdown"
