+++ app/components/checkbox-row/index.ts
@@ -1,0 +1,1 @@
+export { CheckboxRow } from "./checkbox-row"
