+++ app/components/icon-text-button/index.ts
@@ -1,0 +1,1 @@
+export { IconTextButton } from "./icon-text-button"
